Move test files from src to test directory
