+++ packages/react/tsconfig.json
@@ -23,5 +23,5 @@
       "@yorkie-js/sdk/src/*": ["../../packages/sdk/src/*"]
     }
   },
-  "include": ["src"]
+  "include": ["src", "test"]
 }
